+++ cli/main.rs
@@ -866,7 +866,7 @@ async fn run_command(flags: Flags, script: String) -> Result<(), AnyError> {
 
   let mut maybe_coverage_collector =
     if let Some(ref coverage_dir) = program_state.coverage_dir {
-      let session = worker.create_inspector_session();
+      let session = worker.create_inspector_session().await;
 
       let coverage_dir = PathBuf::from(coverage_dir);
       let mut coverage_collector =
